curiosity-agent/requirements.txt
anthropic>=0.40.0
aiohttp>=3.9.0
aiofiles>=23.2.1
pyyaml>=6.0.1

# speech
openai-whisper>=20231117
edge-tts>=6.1.9
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# object recognition (local fallback)
ultralytics>=8.0.200
pillow>=10.0.0

# e-ink display (Waveshare)
RPi.GPIO>=0.7.1; platform_machine=="armv7l" or platform_machine=="aarch64"
spidev>=3.6; platform_machine=="armv7l" or platform_machine=="aarch64"
# waveshare-epaper must be installed manually from vendor SDK

# storage
aiosqlite>=0.20.0

# utilities
python-dotenv>=1.0.0
loguru>=0.7.2
